refactor/#28: userpost 패키지 변경 및 councilpost errorCode 통합
