+++ src/main/java/com/campus/campus/domain/councilpost/application/exception/MajorNotSetException.java
@@ -1,4 +1,4 @@
-package com.campus.campus.domain.userpost.application.exception;
+package com.campus.campus.domain.councilpost.application.exception;
 
 import com.campus.campus.global.common.exception.ApplicationException;
