+++ src/main/java/com/campus/campus/domain/councilpost/policy/PostAccessPolicy.java
@@ -1,10 +1,10 @@
-package com.campus.campus.domain.userpost.policy;
+package com.campus.campus.domain.councilpost.policy;
 
 import org.springframework.stereotype.Component;
 
 import com.campus.campus.domain.council.domain.entity.StudentCouncil;
 import com.campus.campus.domain.user.domain.entity.User;
-import com.campus.campus.domain.userpost.application.exception.PostAccessDeniedException;
+import com.campus.campus.domain.councilpost.application.exception.PostAccessDeniedException;
 
 @Component
 public class PostAccessPolicy {
